Fixed Header File Issue, Restructured Chip8.h and Chip8.cpp
