+++ Chip8Redo/Chip8Redo.pro
@@ -9,6 +9,7 @@ CONFIG += c++17
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
+    ApplicationLoop.cpp \
     Chip8.cpp \
     bindkeys.cpp \
     keybinds.cpp \
@@ -16,6 +17,8 @@ SOURCES += \
     mainwindow.cpp
 
 HEADERS += \
+    ApplicationLoop.h \
+    Chip8.h \
     bindkeys.h \
     keybinds.h \
     mainwindow.h
